fix: codegen'd rule unit should override reflective bind
