+++ submarine-codegen/src/main/java/org/kie/submarine/codegen/rules/RuleUnitInstanceSourceClass.java
@@ -15,16 +15,23 @@
 
 package org.kie.submarine.codegen.rules;
 
+import java.lang.reflect.Method;
+
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.Modifier;
 import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
+import com.github.javaparser.ast.body.MethodDeclaration;
+import com.github.javaparser.ast.expr.CastExpr;
+import com.github.javaparser.ast.expr.EnclosedExpr;
 import com.github.javaparser.ast.expr.MethodCallExpr;
+import com.github.javaparser.ast.expr.MethodReferenceExpr;
 import com.github.javaparser.ast.expr.NameExpr;
 import com.github.javaparser.ast.stmt.BlockStmt;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
-import org.drools.compiler.compiler.io.memory.MemoryFileSystem;
 import org.kie.api.runtime.KieSession;
+import org.kie.submarine.rules.DataSource;
 import org.kie.submarine.rules.impl.AbstractRuleUnitInstance;
+import org.kie.submarine.rules.impl.ListDataSource;
 
 public class RuleUnitInstanceSourceClass {
 
@@ -62,10 +69,56 @@ public class RuleUnitInstanceSourceClass {
         return compilationUnit;
     }
 
+    private MethodDeclaration bindMethod() {
+        // we are currently relying on reflection, but proper way to do this
+        // would be to use JavaParser on the src class AND fallback
+        // on reflection if the class is not available.
+        Class<?> typeClass;
+        try {
+            typeClass = Thread.currentThread().getContextClassLoader().loadClass(canonicalName);
+        } catch (ClassNotFoundException e) {
+            throw new Error(e);
+        }
+
+        MethodDeclaration methodDeclaration = new MethodDeclaration();
+
+        BlockStmt methodBlock = new BlockStmt();
+        methodDeclaration.setName("bind")
+                .addModifier(Modifier.Keyword.PROTECTED)
+                .addParameter(KieSession.class.getCanonicalName(), "rt")
+                .addParameter(typeName, "value")
+                .setType(void.class)
+                .setBody(methodBlock);
+
+        try {
+
+            for (Method m : typeClass.getDeclaredMethods()) {
+                m.setAccessible(true);
+                if (m.getReturnType() == DataSource.class) {
+                    EnclosedExpr casted = new EnclosedExpr(
+                            //  ((ListDataSource) value.$method())
+                            new CastExpr()
+                                    .setType(ListDataSource.class.getCanonicalName())
+                                    .setExpression(new MethodCallExpr(new NameExpr("value"), m.getName())));
+
+                    // .drainInto(rt::insert)
+                    MethodCallExpr drainInto = new MethodCallExpr(casted, "drainInto").addArgument(
+                            new MethodReferenceExpr().setScope(new NameExpr("rt")).setIdentifier("insert"));
+
+                    methodBlock.addStatement(drainInto);
+                }
+            }
+        } catch (Exception e) {
+            throw new Error(e);
+        }
+
+        return methodDeclaration;
+    }
+
     public ClassOrInterfaceDeclaration classDeclaration() {
         ClassOrInterfaceDeclaration classDecl = new ClassOrInterfaceDeclaration()
                 .setName(targetTypeName)
-                .addModifier( Modifier.Keyword.PUBLIC);
+                .addModifier(Modifier.Keyword.PUBLIC);
         classDecl
                 .addExtendedType(
                         new ClassOrInterfaceType(null, AbstractRuleUnitInstance.class.getCanonicalName())
@@ -80,6 +133,7 @@ public class RuleUnitInstanceSourceClass {
                         new NameExpr("value"),
                         new NameExpr("session")
                 )));
+        classDecl.addMember(bindMethod());
         return classDecl;
     }
 }
